review updates:
- documentation
- simplified any_is_complex check
- move is_distributed
