+++ include/ginkgo/core/base/precision_dispatch.hpp
@@ -338,6 +338,31 @@ void mixed_precision_dispatch_real_complex(Function fn, const LinOp* in,
 namespace distributed {
 
 
+/**
+ * Convert the given LinOp from distributed::Vector<...> to
+ * distributed::Vector<ValueType>. The conversion tries to convert the input
+ * LinOp to all Dense types with value type recursively reachable by
+ * next_precision<...> starting from the ValueType template parameter. This
+ * means that all real-to-real and complex-to-complex conversions for default
+ * precisions are being considered. If the input matrix is non-const, the
+ * contents of the modified converted object will be converted back to the input
+ * matrix when the returned object is destroyed. This may lead to a loss of
+ * precision!
+ *
+ * @param matrix  the input matrix which is supposed to be converted. It is
+ *                wrapped unchanged if it is already of type
+ *                distributed::Vector<ValueType>, otherwise it will be converted
+ *                to this type if possible.
+ *
+ * @returns  a detail::temporary_conversion pointing to the (potentially
+ *           converted) object.
+ *
+ * @throws NotSupported  if the input matrix cannot be converted to
+ *                       distributed::Vector<ValueType>
+ *
+ * @tparam ValueType  the value type into whose associated distributed::Vector
+ *                    type to convert the input LinOp.
+ */
 template <typename ValueType>
 detail::temporary_conversion<distributed::Vector<ValueType>>
 make_temporary_conversion(LinOp* matrix)
@@ -351,6 +376,9 @@ make_temporary_conversion(LinOp* matrix)
 }
 
 
+/**
+ * @copydoc make_temporary_conversion
+ */
 template <typename ValueType>
 detail::temporary_conversion<const distributed::Vector<ValueType>>
 make_temporary_conversion(const LinOp* matrix)
@@ -366,6 +394,20 @@ make_temporary_conversion(const LinOp* matrix)
 }
 
 
+/**
+ * Calls the given function with each given argument LinOp temporarily
+ * converted into distributed::Vector<ValueType> as parameters.
+ *
+ * @param fn  the given function. It will be passed one (potentially const)
+ *            distributed::Vector<ValueType>* parameter per parameter in the
+ *            parameter pack `linops`.
+ * @param linops  the given arguments to be converted and passed on to fn.
+ *
+ * @tparam ValueType  the value type to use for the parameters of `fn`.
+ * @tparam Function  the function pointer, lambda or other functor type to call
+ *                   with the converted arguments.
+ * @tparam Args  the argument type list.
+ */
 template <typename ValueType, typename Function, typename... Args>
 void precision_dispatch(Function fn, Args*... linops)
 {
@@ -373,6 +415,15 @@ void precision_dispatch(Function fn, Args*... linops)
 }
 
 
+/**
+ * Calls the given function with the given LinOps temporarily converted to
+ * distributed::Vector<ValueType>* as parameters.
+ * If ValueType is real and both input vectors are complex, uses
+ * distributed::Vector::get_real_view() to convert them into real matrices after
+ * precision conversion.
+ *
+ * @see precision_dispatch()
+ */
 template <typename ValueType, typename Function>
 void precision_dispatch_real_complex(Function fn, const LinOp* in, LinOp* out)
 {
@@ -396,6 +447,9 @@ void precision_dispatch_real_complex(Function fn, const LinOp* in, LinOp* out)
 }
 
 
+/**
+ * @copydoc precision_dispatch_real_complex(Function, const LinOp*, LinOp*)
+ */
 template <typename ValueType, typename Function>
 void precision_dispatch_real_complex(Function fn, const LinOp* alpha,
                                      const LinOp* in, LinOp* out)
@@ -424,6 +478,9 @@ void precision_dispatch_real_complex(Function fn, const LinOp* alpha,
 }
 
 
+/**
+ * @copydoc precision_dispatch_real_complex(Function, const LinOp*, LinOp*)
+ */
 template <typename ValueType, typename Function>
 void precision_dispatch_real_complex(Function fn, const LinOp* alpha,
                                      const LinOp* in, const LinOp* beta,
@@ -459,6 +516,18 @@ void precision_dispatch_real_complex(Function fn, const LinOp* alpha,
 }  // namespace distributed
 
 
+/**
+ * Calls the given function with the given LinOps temporarily converted to
+ * either distributed::Vector<ValueType>* or matrix::Dense<ValueType> as
+ * parameters. The choice depends on the runtime type of `in` and `out` is
+ * assumed to fall into the same category.
+ * If ValueType is real and both input vectors are complex, uses
+ * distributed::Vector::get_real_view(), or matrix::Dense::get_real_view() to
+ * convert them into real matrices after precision conversion.
+ *
+ * @see precision_dispatch()
+ * @see distributed::precision_dispatch()
+ */
 template <typename ValueType, typename Function>
 void precision_dispatch_real_complex_distributed(Function fn, const LinOp* in,
                                                  LinOp* out)
@@ -471,6 +540,10 @@ void precision_dispatch_real_complex_distributed(Function fn, const LinOp* in,
 }
 
 
+/**
+ * @copydoc precision_dispatch_real_complex_distributed(Function, const LinOp*,
+ * LinOp*)
+ */
 template <typename ValueType, typename Function>
 void precision_dispatch_real_complex_distributed(Function fn,
                                                  const LinOp* alpha,
@@ -485,6 +558,10 @@ void precision_dispatch_real_complex_distributed(Function fn,
 }
 
 
+/**
+ * @copydoc precision_dispatch_real_complex_distributed(Function, const LinOp*,
+ * LinOp*)
+ */
 template <typename ValueType, typename Function>
 void precision_dispatch_real_complex_distributed(Function fn,
                                                  const LinOp* alpha,
@@ -504,6 +581,15 @@ void precision_dispatch_real_complex_distributed(Function fn,
 #else
 
 
+/**
+ * Calls the given function with the given LinOps temporarily converted to
+ * matrix::Dense<ValueType> as parameters.
+ * If ValueType is real and both input vectors are complex, uses
+ * distributed::Vector::get_real_view(), or matrix::Dense::get_real_view() to
+ * convert them into real matrices after precision conversion.
+ *
+ * @see precision_dispatch()
+ */
 template <typename ValueType, typename Function, typename... Args>
 void precision_dispatch_real_complex_distributed(Function fn, Args*... args)
 {
